Diagnostic script to be more useful
Repurpose the aging diagnostic script to allow to print better
information about a users running environment. With hints for the
more obvious discovered issues (moose-dev not in sync).

Also includes the ability to build everything from source as a
"proof of concept" that the machine in question is sound enough
to build/run MOOSE.

Closes #26843